+++ packages/e2e-tests/src/deploy.ts
@@ -57,6 +57,7 @@ const VKEY = JSON.parse(fs.readFileSync(VKEY_PATH).toString());
 
 export interface NocturneDeployArgs {
   screeners: Address[];
+  subtreeBatchFillers: Address[];
 }
 
 export interface TestActorsConfig {
@@ -164,6 +165,7 @@ export async function setupTestDeployment(
     KEYS_TO_WALLETS(provider);
   const contractDeployment = await deployContractsWithDummyAdmins(deployerEoa, {
     screeners: [screenerEoa.address],
+    subtreeBatchFillers: [deployerEoa.address, subtreeUpdaterEoa.address],
   });
 
   await checkNocturneContractDeployment(
@@ -205,6 +207,7 @@ export async function setupTestDeployment(
     proms.push(startBundler(bundlerConfig));
   }
 
+  // deploy subtree updater if requested
   let subtreeUpdaterContainer: Dockerode.Container | undefined;
   if (config.include.subtreeUpdater) {
     const givenSubtreeUpdaterConfig = config.configs?.subtreeUpdater ?? {};
@@ -310,6 +313,7 @@ export async function deployContractsWithDummyAdmins(
       walletOwner: "0x3CACa7b48D0573D793d3b0279b5F0029180E83b6",
       depositManagerOwner: "0x3CACa7b48D0573D793d3b0279b5F0029180E83b6",
       screeners: args.screeners,
+      subtreeBatchFillers: args.subtreeBatchFillers,
     },
     {
       useMockSubtreeUpdateVerifier: